+++ lib/index.ts
@@ -1,4 +1,5 @@
 export {exitProcessWithError} from './exitProcessWithError'
+export {Package} from './Package'
 export {readFile} from './readFile'
 export {rename} from './rename'
 export {writeFile} from './writeFile'